server module now handles POST requests
Big addendum: TIL that POST payloads MAY be sent along with headers, or in a separate chunk.
If we continue to examine packets statelessly, we cant determine a payload is part of any specific POST request if it arrives separately.
A very simple workaround is used - this will need to change in the future.